chore(engine-rest): move password policy endpoints to identity service
* move password policy REST endpoints to identity REST service
* respond with 404 on GET /password-policy if no policy is configured

Related to CAM-9936
