+++ engine-rest/engine-rest/src/main/java/org/camunda/bpm/engine/rest/impl/DefaultProcessEngineRestServiceImpl.java
@@ -41,7 +41,6 @@ import org.camunda.bpm.engine.rest.MessageRestService;
 import org.camunda.bpm.engine.rest.MetricsRestService;
 import org.camunda.bpm.engine.rest.MigrationRestService;
 import org.camunda.bpm.engine.rest.ModificationRestService;
-import org.camunda.bpm.engine.rest.PasswordPolicyRestService;
 import org.camunda.bpm.engine.rest.ProcessDefinitionRestService;
 import org.camunda.bpm.engine.rest.ProcessInstanceRestService;
 import org.camunda.bpm.engine.rest.SignalRestService;
@@ -211,11 +210,6 @@ public class DefaultProcessEngineRestServiceImpl extends AbstractProcessEngineRe
   @Path(VersionRestService.PATH)
   public VersionRestService getVersionRestService() {
     return super.getVersionRestService(null);
-  }
-
-  @Path(PasswordPolicyRestService.PATH)
-  public PasswordPolicyRestService getPasswordPolicyRestService() {
-    return super.getPasswordPolicyRestService(null);
   }
 
   @Override
